web-app/backend/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
flask-restplus==0.13.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.15.6
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
